add numerical checking to run-test.sh; pipe diagnostic messages to stdout so you can use tee

+++ jfpjc/testbench/jfpjc_images_test/run_test.sh
@@ -2,7 +2,7 @@
 
 # If we don't have images yet, get them.
 IMAGE_COUNT=$(ls -1q ./dsp-test-images/ | wc -l)
-echo $IMAGE_COUNT " images found"
+>&2 echo $IMAGE_COUNT " images found"
 if [ $IMAGE_COUNT -lt 20 ]
 then
     echo "downloading dsp-test-images set from github"
@@ -14,8 +14,8 @@ then
 fi
 
 IMAGE_COUNT=$(ls -1q ./dsp-test-images/ | wc -l)
-echo $IMAGE_COUNT " images downloaded"
-echo
+>&2 echo $IMAGE_COUNT " images downloaded"
+>&2 echo
 
 PROJECT_BASE=$(git rev-parse --show-toplevel)
 
@@ -23,10 +23,13 @@ for f in ./dsp-test-images/*.tiff; do
     rm output.jpg 2> /dev/null
 
     # For all of the images, convert them to 320x240 grayscale pgm for comparison purposes.
-    # echo "converting " $f
+    >&2 echo "converting " $f
     newpgm=$(basename ${f%.tiff})
     convert "$f" "${newpgm}.pgm"
-    convert -gravity center -crop 320x240+0+0 "${newpgm}.pgm" "${newpgm}_320x240.pgm"
+    convert -gravity center -crop 320x240+0+0 "${newpgm}.pgm" "${newpgm}_crop.pgm"
+    convert "${newpgm}_crop.pgm" -extent 320x240 -gravity NorthWest -background black "${newpgm}_320x240.pgm"
+    rm "${newpgm}.pgm"
+    rm "${newpgm}_crop.pgm"
 
     # Convert the image to a hex file
     $PROJECT_BASE/tools/image_to_hex.py "${newpgm}_320x240.pgm" > testimg.hex
@@ -36,22 +39,27 @@ for f in ./dsp-test-images/*.tiff; do
     ./jfpjc_tb.vvp
 
     # Compare input and output image, providing a similarity score
-    echo "Image ${newpgm}.pgm compressed to size" $(ls -l "output.jpg"  | awk '{print $5}')
-    echo "output.jpg difference score to baseline"
-    $PROJECT_BASE/tools/image_error.py "${newpgm}_320x240.pgm" "output.jpg"
-    if [ "$?" -ne 0 ]; then exit 1; fi
-    echo "imagemagick difference score to baseline"
     convert -quality 100 "${newpgm}_320x240.pgm" "${newpgm}_320x240_q100.jpg"
-    $PROJECT_BASE/tools/image_error.py "${newpgm}_320x240.pgm" "${newpgm}_320x240_q100.jpg"
+    echo "Image ${newpgm}.pgm compressed to size" $(ls -l "output.jpg"  | awk '{print $5}')
+    SCORE1=$($PROJECT_BASE/tools/image_error.py "${newpgm}_320x240.pgm" "output.jpg")
     if [ "$?" -ne 0 ]; then exit 1; fi
-    echo
-    echo "================================================================"
-    echo
+    SCORE2=$($PROJECT_BASE/tools/image_error.py "${newpgm}_320x240.pgm" "${newpgm}_320x240_q100.jpg")
+    if [ "$?" -ne 0 ]; then exit 1; fi
+    >&2 echo "     jfpjc     imagemagick"
+    printf "% 10.1f      % 10.1f" $SCORE1 $SCORE2
 
-    rm output.jpg 2> /dev/null
-    rm "${newpgm}.pgm"
-    rm "${newpgm}_320x240.pgm"
-    rm "${newpgm}_320x240_q100.jpg"
+    THRESHOLD=2000.0
+    if (($(echo "sqrt(($SCORE1 - $SCORE2) * ($SCORE1 - $SCORE2)) > $THRESHOLD" | bc -l))); then
+        >&2 printf " ************************"
+        stdbuf -i0 -o0 -e0 echo
+        >&2 echo "Scores for $f differ by more than $THRESHOLD. Retaining images."
+        mv "output.jpg" "${newpgm}_320x240_jfpjc.jpg"
+    else
+        stdbuf -i0 -o0 -e0 echo
+        rm output.jpg
+        rm "${newpgm}_320x240.pgm"
+        rm "${newpgm}_320x240_q100.jpg"
+    fi
 
     # Also make a 100 and 90 quality jpeg conversion of the pnm images for
 done
